feat(v1): rename SKX harness
