+++ verifiers/v1/harnesses/__init__.py
@@ -9,10 +9,10 @@ from verifiers.v1.harnesses.mini_swe_agent import (
     MiniSWEAgentHarness,
     MiniSWEAgentHarnessConfig,
 )
-from verifiers.v1.harnesses.null import NullHarness, NullHarnessConfig
 from verifiers.v1.harnesses.pi import PiHarness, PiHarnessConfig
 from verifiers.v1.harnesses.pool import PoolHarness, PoolHarnessConfig
 from verifiers.v1.harnesses.rlm import RLMHarness, RLMHarnessConfig
+from verifiers.v1.harnesses.skx import SkxHarness, SkxHarnessConfig
 from verifiers.v1.harnesses.terminus_2 import Terminus2Harness, Terminus2HarnessConfig
 
 __all__ = [
@@ -26,14 +26,14 @@ __all__ = [
     "KimiCodeHarnessConfig",
     "MiniSWEAgentHarness",
     "MiniSWEAgentHarnessConfig",
-    "NullHarness",
-    "NullHarnessConfig",
     "PiHarness",
     "PiHarnessConfig",
     "PoolHarness",
     "PoolHarnessConfig",
     "RLMHarness",
     "RLMHarnessConfig",
+    "SkxHarness",
+    "SkxHarnessConfig",
     "Terminus2Harness",
     "Terminus2HarnessConfig",
 ]
