+++ verifiers/v1/harnesses/null/__init__.py
@@ -1,3 +1,8 @@
-from verifiers.v1.harnesses.null.harness import NullHarness, NullHarnessConfig
+"""Compatibility alias for historical configs; new SKX runs use ``id = "skx"``."""
+
+from verifiers.v1.harnesses.skx import SkxHarness, SkxHarnessConfig
+
+NullHarness = SkxHarness
+NullHarnessConfig = SkxHarnessConfig
 
 __all__ = ["NullHarness", "NullHarnessConfig"]
